make communication protocol complient with XCSoar and fix some bugs
1. introduced support for POV,C,RPO,... and POV,C,IPO,.... consitent
with documentation

2. leave POV,C,POL,... as undocumented feature; should be eliminated lateron

3. prefer Real Polar information (POV,C,RPO,...) over Ideal Polar with
later calculation of real polar. The overload information from XCSoar is not
correct if refernce_mass != dry_mass . Therefor the information is not
enough to calculate real polar from ideal polar. And XCSoar does it all
for as and provides it ready to use in POV,C,RPO,...

4. parse_NMEA_command() in variod.c porcessed the incomming sentences in
reverse order if several sentences came with one recv().