initial release of testbench tools to variod
It provides sensord_mock which can act on behalf of sensord.

It provides my_netcat which is the equivalent of netcat "nc -l"
for systems, where "nc -l" is not available. It can be used to
act on behalf of XCSoar to see what it would send to it.

Please consult README.txt or more detail.