addressing issues listed in #20
It improves robustness of NMEA parsers.
It presents one NMEA sentence at a time to be parsed.
It follows the guidelines for NMEA sentence structure given here:
https://en.wikipedia.org/wiki/NMEA_0183#Message_structure
with listed exeptions:
NMEA sentences are only valid if they have a checksum.
The <CR><LF> at the end of the sentence is not enforced.
One of the 2 is sufficient, but not less than 1 is required.

The next steps:
Implement the request to XCSoar to send current settings.
Handle incomplete senteces from recv().
Both are marked in the code as TODO: